requirements.txt
PyYaml
tabula-py
tabulate
jpype1
requests
boto3